tests/requirements.txt
# Test requirements for SLURM Ansible modules
# These packages are required to run the test suite

# Core testing framework
pytest>=6.0.0
pytest-cov>=2.10.0
pytest-mock>=3.0.0

# Ansible testing
ansible-core>=2.12.0
ansible-test

# Code quality
flake8>=3.8.0
pylint>=2.6.0
black>=21.0.0

# Coverage reporting
coverage>=5.0.0

# Mock and testing utilities
mock>=4.0.0
responses>=0.12.0

# Documentation
sphinx>=3.0.0
sphinx-rtd-theme>=0.5.0 